# server/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HEXUDON Server API Reference
  description: Tài liệu OpenAPI 3.1 cho máy chủ HEXUDON
  version: 0.1.0
servers:
  - url: /
...
paths:
...
  /api/game/config:
    get:
      tags:
        - Config
      summary: Cấu hình bảng một lần cho đội gọi yêu cầu (CHỈ ĐỘI)
      description: Đọc cấu hình trận đấu ban đầu dành riêng cho đội thi đấu (yêu cầu xác thực token).
      security:
        - BearerAuth: []
      parameters:
        - name: game_id
          in: query
          required: true
          description: ID của trận đấu
          schema:
            type: string
      responses:
        '200':
          description: Phản hồi thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchConfigResponse'
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Trận đấu không tồn tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/game/agent-types:
    post:
      tags:
        - Play
      summary: Chọn loại Agent trước trận đấu (CHỈ ĐỘI)
      description: Chọn loại tuần tra (patrol) hoặc nạp nhiên liệu (refuel) cho từng Agent của đội.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelectAgentTypesRequest'
      responses:
        '204':
          description: Chọn loại Agent thành công, không có nội dung trả về
        '401':
          description: Chưa xác thực
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Yêu cầu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SpotResponse:
      type: object
      required:
        - brand
        - pos
        - stocks
      properties:
        brand:
          type: integer
        pos:
          type: integer
        stocks:
          type: integer
...
    MatchConfigResponse:
      type: object
      description: Cấu hình trận đấu trả về cho đội
      required:
        - startsAt
        - daySeconds
        - daySteps
        - map
        - spots
        - agents
        - fuelLimits
        - players
        - busyThreshold
        - jammedThreshold
      properties:
        startsAt:
          type: integer
          format: int64
        daySeconds:
          type: array
          items:
            type: number
            format: double
        daySteps:
          type: array
          items:
            type: integer
        map:
          $ref: '#/components/schemas/MapResponse'
        spots:
          type: array
          items:
            $ref: '#/components/schemas/SpotResponse'
        agents:
          type: array
          items:
            type: integer
        fuelLimits:
          type: integer
        players:
          type: integer
        busyThreshold:
          type: number
          format: double
        jammedThreshold:
          type: number
          format: double
...
    MapResponse:
      type: object
      required:
        - height
        - width
        - cells
      properties:
        height:
          type: integer
        width:
          type: integer
        cells:
          type: array
          items:
            type: array
            items:
              type: integer
    SelectAgentTypesRequest:
      type: object
      description: Yêu cầu chọn loại Agent
      required:
        - game_id
        - types
      properties:
        game_id:
          type: string
        types:
          type: array
          items:
            type: integer
          description: Danh sách loại cho từng Agent (0=patrol, 1=refuel)
...
    ValidationErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        rejectedValue:
          type: string
          nullable: true
        message:
          type: string
    ErrorResponse:
      type: object
      required:
        - errorCode
        - message
        - timestamp
      properties:
        errorCode:
          type: string
        message:
          type: string
        timestamp:
          type: integer
          format: int64
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
...